Fix safe navigation operator description in conditionals.md
Corrected the explanation of the safe navigation operator to indicate it returns 'null' instead of 'undefined'. Undefined is not defined in the book, but seems to be a mistake. 'null' may be a problem too, if CF does not have null enabled. This section is troublesome to the reader, in any case.